Partial changes towards daijin configure
